Add files via upload
The topic of your EECS 3482 research project was:   Australian National University Hack/Breach of 2018.
Objective:
The goal of this research project is to give the students an opportunity to independently investigate a
recent real-world case of data breach and/or hack using the knowledge acquired in the course. In
particular, the student is expected to find and research various online sources in order to gain in-depth
understanding of the given breach/hack, including: when and how the breach/hack occurred, who the
main stakeholders of the breach/hack were, what the most significant consequences and lessons of the
breach/hack are, etc. The findings of the research will be submitted for evaluation in the form of a written
report.